refactor grab_unsolved_problems into a class function
This makes it easier to write the base function, as the numbered flag can just be a class var.
I also realized that I could write the if-else statement differently... Yay, reducing nesting.
I had the new fn write the pool to the class var instead of doing that after calling it.