refactor grab_list_of_problems into a class function
And because it does not interact with any of the class variables, it can be the first time I've declared `@staticmethod`, lol.
I also rewrote it to use try-except instead of an if catch clause.
I also rewrote it to use list comprehension instead of a loop.

I updated `pool` to be `pool_of_problems`

I also uppated `__init__.py` because I forgot to before.